backend/requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
pgvector
google-generativeai
python-dotenv
yfinance
pandas
pydantic
pydantic-settings
requests
